d3d9-nine: use 'd3d9' as basename for the spec file
The basename is used by `winebuild` as a filename, and we need to use
'd3d9' since we substitute that library.

Before this patch, WINE would load 'd3d9-nine.dll' and add it to its
'InLoadOrderModuleList' under that basename. That in return makes
LdrGetDllHandle()/GetModuleHandle*() calls for 'd3d9.dll' not find our
substitute in that list and it walks PATH all over again.

GetModuleHandle*("d3d9.dll") is used _a lot_ on the Stalker games, and
this patch brings down the time to load a savegame from >2 minutes to
<15 seconds.