travis: packaging fixes
Use same revision number for stable releases as for unstable ones.
This is required as 1 was used for testing and tag 0.3 won't cause
an update on repo pull as its lower than 1.

Add the tag as description and use newline on defines to fix possible
build error.
Fix packaging conflict with itself.

Signed-off-by: Patrick Rudolph <siro@das-labor.org>