switch CI to winehq 6.0 packages
The current releases are binary incompatible with WINE v8.3, as that version
removed libwine, which all our releases link against.

We need to build with >=v5.7 in order to drop that dependency from the
resulting binaries.

Hence switch to the old stable v6.0 to build our releases. That fixes the
binary compatibility with >=v8.3, but unfortunately limits it to >=v5.7 at the
same time.